requirements.txt
APScheduler==3.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==4.2.2
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.1
greenlet==3.0.3
idna==3.7
lxml==5.2.1
platformdirs==4.2.0
python-dotenv==1.1.1
python-telegram-bot==13.7
pytz==2025.2
requests==2.32.5
requests-cache==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
tornado==6.5.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2025.2
tzlocal==5.3.1
url-normalize==1.4.3
urllib3==2.5.0
